Update Adv02.py
Cosmetic updates.

+++ FunctionMachines/Adv02.py
@@ -13,6 +13,18 @@ import os
 import os.path
 import json
 
+def create_filename(key):
+    return f"{key.lower()}.json"
+
+
+def base_name(file):
+    return f'Record: {file.replace(".json", "")}'
+
+
+def should_delete(file):
+    return input(f"Delete existing '{file}'? [Y/N] ") == 'Y'
+
+
 def list_files():
     results = []
     for file in os.listdir("."):
@@ -21,11 +33,6 @@ def list_files():
         results.append(file)
     return results
 
-def create_filename(key):
-    return f"{key.lower()}.json"
-
-def should_delete(file):
-    return input(f"Delete existing '{file}'? [Y/N] ") == 'Y'
 
 def create():
     # NOTE: OrderedDict best before 3.7
@@ -36,9 +43,12 @@ def create():
     for key in record:
         record[key] = input(f"Enter {key}: ")
     file = create_filename(record['Name'])
+    if os.path.exists(file) and not should_delete(file):
+        return False
     with open(file, "w") as fh:
         json.dump(record, fh)
     return True
+
 
 def read():
     ''' Scenario: 
@@ -48,25 +58,27 @@ def read():
     name = input("Enter Name: ")
     file = create_filename(name)
     if not os.path.exists(file):
-        print(f"File '{file} not found.")
+        print(f"File '{file}' not found.")
         return None
     with open(file) as fh:
         record = json.load(fh)
-        print(record)
-        return record
+        if record:
+            print(record)
+            return record
     return None
+
 
 def update():
     ''' Scenario:
     User selects & update a record.
-    Returns True if updated, else False.
+    Returns True if updated, else False
     '''
     record = read()
     if record:
         changed = False
         for key in record:
             print(record[key])
-            value = input("Update: (enter to keep) ")
+            value = input("Update: (enter to keep): ")
             if value:
                 record[key] = value
                 changed = True
@@ -76,6 +88,7 @@ def update():
     with open(file, "w") as fh:
         json.dump(record, fh)
     return True
+
 
 def delete():
     ''' Scenario:
@@ -88,12 +101,13 @@ def delete():
         if should_delete(file):
             os.unlink(file)
 
+
 def search():
     ''' Scenario:
     User searches for a string, in all records.
     Results are displayed.
     Returns number of occurrences found.
-    Returns 0 on error, or nothing, was found.
+    Returns 0 on error, or nothing was found.
     '''
     search = input("Locate: ")
     count = 0
@@ -101,30 +115,30 @@ def search():
         for file in list_files():
             with open(file) as fh:
                 record = json.load(fh)
-                if record:
-                    for key in record:
-                        if record[key].find(search) != -1:
-                            report = file.replace(".json", "")
-                            print(
-                                f'Record: {report}\t', 
-                                  record[key])
-                            count += 1
+                if not record:
+                    continue
+                for key in record:
+                    if record[key].find(search) != -1:
+                        report = base_name(file)
+                        print(report, 
+                              record[key], sep='\t')
+                        count += 1
     return count
+
 
 def listf():
     ''' Scenario:
     Display the key (file) names, if any.
     Always returns None.
     '''
-    for file in os.listdir("."):
-        if not file.endswith(".json"):
-            continue
-        print(f"{file.replace('.json','')}")
+    for file in list_files():
+        print("\t", base_name(file))
+
 
 if __name__ == '__main__':
     options = {
         'c':create, 'r':read, 'u':update, 'd':delete,
-        's':search, 'l':listf, 'q':quit
+        'l':listf, 's':search, 'q':quit
         }
     while True:
         op = input("Option: ")
